move repl_commands.rs to debug_adapter::dap
