+++ debugger/src/debug_adapter/dap/adapter.rs
@@ -1,4 +1,4 @@
-use crate::debug_adapter::repl_commands::{build_expanded_commands, command_completions};
+use crate::debug_adapter::dap::repl_commands::{build_expanded_commands, command_completions};
 use crate::{
     debug_adapter::{dap::dap_types, protocol::ProtocolAdapter},
     debugger::{
@@ -26,7 +26,7 @@ use serde::{de::DeserializeOwned, Serialize};
 use std::{convert::TryInto, path::Path, str, string::ToString, time::Duration};
 
 use super::core_status::DapStatus;
-use super::request_helpers::{disassemble_target_memory, get_dap_source};
+use super::request_helpers::{disassemble_target_memory, get_dap_source, get_variable_reference};
 
 /// Progress ID used for progress reporting when the debug adapter protocol is used.
 type ProgressId = i64;
@@ -1906,45 +1906,5 @@ pub fn get_arguments<T: DeserializeOwned, P: ProtocolAdapter>(
                 req.command
             )))
         }
-    }
-}
-
-/// The DAP protocol uses three related values to determine how to invoke the `Variables` request.
-/// This function retrieves that information from the `DebugInfo::VariableCache` and returns it as
-/// (`variable_reference`, `named_child_variables_cnt`, `indexed_child_variables_cnt`)
-pub(crate) fn get_variable_reference(
-    parent_variable: &probe_rs::debug::Variable,
-    cache: &mut probe_rs::debug::VariableCache,
-) -> (i64, i64, i64) {
-    if !parent_variable.is_valid() {
-        return (0, 0, 0);
-    }
-    let mut named_child_variables_cnt = 0;
-    let mut indexed_child_variables_cnt = 0;
-    if let Ok(children) = cache.get_children(Some(parent_variable.variable_key)) {
-        for child_variable in children {
-            if child_variable.is_indexed() {
-                indexed_child_variables_cnt += 1;
-            } else {
-                named_child_variables_cnt += 1;
-            }
-        }
-    };
-
-    if named_child_variables_cnt > 0 || indexed_child_variables_cnt > 0 {
-        (
-            parent_variable.variable_key,
-            named_child_variables_cnt,
-            indexed_child_variables_cnt,
-        )
-    } else if parent_variable.variable_node_type.is_deferred()
-        && parent_variable.get_value(cache) != "()"
-    {
-        // We have not yet cached the children for this reference.
-        // Provide DAP Client with a reference so that it will explicitly ask for children when the user expands it.
-        (parent_variable.variable_key, 0, 0)
-    } else {
-        // Returning 0's allows VSCode DAP Client to behave correctly for frames that have no variables, and variables that have no children.
-        (0, 0, 0)
     }
 }
