+++ debugger/src/debug_adapter/dap/mod.rs
@@ -4,5 +4,12 @@ pub(crate) mod adapter;
 pub(crate) mod core_status;
 /// The MS DAP api (from json spec), and extensions (custom), for communicating with the MS DAP client.
 pub(crate) mod dap_types;
-/// ...
+/// Handle the various "gdb-like" commands that are sent to the debug adapter, from the Debug Console REPL window.
+/// These commands are not part of the DAP protocol, but are implemented by the debug adapter to provide a
+/// gdb-like experience to users who prefer that to the VS Code UX.
+/// It doesn't make sense to implement all gdb commands, and this implementation will focus on the ones that
+/// are most useful to users, and this list is expected to grow over time.
+pub(crate) mod repl_commands;
+/// The logic for handling the various MS DAP requests, implemented so that it can be used by both the
+/// [`adapter::DebugAdapter`] and the [`repl_commands::ReplCommands`].
 pub(crate) mod request_helpers;
